refactor subcommand handling
